Move mremap to cpu-specific modules for Linux
